requirements.txt
selenium==4.1.3
pandas==1.4.1
lxml==4.8.0
mysql-connector-python==8.0.28
python-dotenv==0.19.2